+++ Rectangles/Rectangles/Square.h
@@ -10,6 +10,6 @@
 
 @interface Square: Rectangle
 
--(void) setSide: (int) s; 
--(int) side;
+-(void) setSide: (float) s; 
+-(float) side;
 @end
