+++ Rectangles/Rectangles/Square.m
@@ -10,12 +10,12 @@
 
 @implementation Square: Rectangle
 
--(void) setSide: (int) s 
+-(void) setSide: (float) s 
 {
 	[self setWidth: s andHeight: s]; 
 }
 
--(int) side
+-(float) side
 {
 	return self.width;
 }
